+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
     * Added `select()` and `unselect()` methods to `CollectionInterface` and its base classes
     * Added `orderBy()` and `limit()` methods to `ObjectCollectionInterface` and its base classes
+    * Flex: Added support for custom object index classes
 1. [](#improved)
     * Improve Flex storage
 1. [](#bugfix)